Clarify GitHub docs
Added more detail on the authentication, configuration and usage of
`code-review` when used against a private GitHub enterprise instance. The
information is taken from setting up the package myself, and
also from [this
issue](https://github.com/wandersoncferreira/code-review/issues/174).

Mentions of a bug when using `code-review-start` are removed because the
function does work in my setup, so the doc seems outdated in that respect.